+++ servers/seriesHTTPhandler.go
@@ -3,6 +3,7 @@ package servers
 import (
 	"encoding/json"
 	"fmt"
+	"gateserver/spaces"
 	"gateserver/storage"
 	"gateserver/support"
 	"log"
@@ -33,8 +34,8 @@ func seriesHTTPhandler() http.Handler {
 		defer func() {
 			if e := recover(); e != nil {
 				go func() {
-					support.DLog <- support.DevData{"servers,seriesHTTPhandler",
-						support.Timestamp(), "", []int{1}, true}
+					support.DLog <- support.DevData{"servers.seriesHTTPhandler",
+						support.Timestamp(), "handle crashed", []int{1}, true}
 				}()
 				log.Println("servers.seriesHTTPhandler: recovering from: ", e)
 			}
@@ -57,9 +58,9 @@ func seriesHTTPhandler() http.Handler {
 			}
 		}
 		dbgMutex.RUnlock()
-		if authorised {
-			log.Println("servers.seriesHTTPhandler: answering (debug) authorised device at address:", ip)
-		}
+		// if authorised {
+		// 	log.Println("servers.seriesHTTPhandler: answering (debug) authorised device at address:", ip)
+		// }
 		//fmt.Println(authorised, ip)
 
 		params := make(map[string]string)
@@ -81,7 +82,7 @@ func seriesHTTPhandler() http.Handler {
 			}
 		}
 
-		//fmt.Println(params)
+		// fmt.Println(params)
 
 		label := support.StringLimit(params["type"], support.LabelLength)
 		label += support.StringLimit(params["space"], support.LabelLength)
@@ -113,7 +114,7 @@ func seriesHTTPhandler() http.Handler {
 			var rt []storage.SampleData
 			if params["start"] != "" && params["end"] != "" {
 
-				// if not in debug mode, the request will not provide data for the current day
+				// if not in authorised via pin, the request will not provide data for the current day
 				if !authorised {
 					if eni, err := strconv.Atoi(params["end"]); err == nil {
 						en := time.Unix(int64(eni/1000), 0)
@@ -129,6 +130,8 @@ func seriesHTTPhandler() http.Handler {
 						_, _ = fmt.Fprintf(w, "")
 						return
 					}
+				} else {
+					log.Println("servers.seriesHTTPhandler: answering (debug) authorised device at address:", ip)
 				}
 
 				var st, en int64
@@ -150,23 +153,77 @@ func seriesHTTPhandler() http.Handler {
 				case "sample":
 					s0 = &storage.SeriesSample{Stag: label, Sts: st}
 					s1 = &storage.SeriesSample{Stag: label, Sts: en}
+					fmt.Println(s0, s1)
 					if tag, ts, vals, e := storage.ReadSeriesTS(s0, s1, params["analysis"] != "current"); e == nil {
 						rt = s0.UnmarshalSliceSS(tag, ts, vals)
 					}
+					// fmt.Println(rt)
 					if e := json.NewEncoder(w).Encode(rt); e != nil {
 						_, _ = fmt.Fprintf(w, "")
 					}
 				case "entry":
 					// TODO how to we handle authorisation, best is to merge sample and entry reporting and report corrupted data in values do not coincides
-					if !authorised {
-						fmt.Println("not authorised")
-						_, _ = fmt.Fprintf(w, "")
-						return
-					}
 					var convertedRt []storage.JsonSeriesEntries
+
+					// if !authorised {
+					// 	fmt.Println("not authorised")
+					// 	// _, _ = fmt.Fprintf(w, "")
+					// 	if e := json.NewEncoder(w).Encode(convertedRt); e != nil {
+					// 		_, _ = fmt.Fprintf(w, "")
+					// 	}
+					// 	return
+					// }
+					// s0 = &storage.SeriesEntries{Stag: label, Sts: st}
+					// s1 = &storage.SeriesEntries{Stag: label, Sts: en}
+					// // fmt.Println(s0, s1)
+					// //os.Exit(1)
+					// if tag, ts, vals, e := storage.ReadSeriesTS(s0, s1, params["analysis"] != "current"); e == nil {
+					// 	rt = s0.UnmarshalSliceSS(tag, ts, vals)
+					// }
+					// for _, v := range rt {
+					// 	convertedTemp := storage.JsonSeriesEntries{}
+					// 	seVal := storage.SeriesEntries{}
+					// 	codedVal := v.Marshal()
+					// 	_ = seVal.Unmarshal(codedVal)
+					// 	seVal.Stag = v.Tag()
+					// 	convertedTemp.ExpandEntries(seVal)
+					// 	convertedRt = append(convertedRt, convertedTemp)
+					// }
+
+					// TODO starts development
+					var dataPeriod int
+					for _, el := range spaces.AvgAnalysis {
+						if el.Name == support.StringLimit(params["analysis"], support.LabelLength) {
+							dataPeriod = el.Interval
+						}
+					}
+					fmt.Println(dataPeriod)
+					// os.Exit(1)
+
+					// add also the reading of samples and compare values ... this added part crashes
+					s0s := &storage.SeriesSample{Stag: strings.Replace(label, "entry___", "sample__", -1), Sts: st}
+					s1s := &storage.SeriesSample{Stag: strings.Replace(label, "entry___", "sample__", -1), Sts: en}
+					// fmt.Println(s0s, s1s)
+					// fmt.Println("1")
+					var referenceSamples []storage.SeriesSample
+					if tag, ts, vals, e := storage.ReadSeriesTS(s0s, s1s, params["analysis"] != "current"); e == nil {
+						// fmt.Println(tag, ts, vals)
+						referenceSamples = s0s.UnmarshalSliceNative(tag, ts, vals)
+					}
+					// fmt.Println("3")
+					// for _, el := range rts {
+					// 	convertedRts = append(convertedRts, storage.SeriesSample{el.Tag(), el.Sts(), el.val})
+					// 	// fmt.Println(el)
+					// }
+					fmt.Println(len(referenceSamples))
+					for _, el := range referenceSamples {
+						fmt.Println(el.Sts, el.Sval)
+					}
+					// os.Exit(1)
+
 					s0 = &storage.SeriesEntries{Stag: label, Sts: st}
 					s1 = &storage.SeriesEntries{Stag: label, Sts: en}
-					//fmt.Println(s0,s1)
+					// fmt.Println(s0, s1)
 					//os.Exit(1)
 					if tag, ts, vals, e := storage.ReadSeriesTS(s0, s1, params["analysis"] != "current"); e == nil {
 						rt = s0.UnmarshalSliceSS(tag, ts, vals)
@@ -180,17 +237,13 @@ func seriesHTTPhandler() http.Handler {
 						convertedTemp.ExpandEntries(seVal)
 						convertedRt = append(convertedRt, convertedTemp)
 					}
-					// TODO development
-					// add also the reading of samples and compare values ...
-					//s0a := &storage.SeriesSample{Stag: label, Sts: st}
-					//s1a := &storage.SeriesSample{Stag: label, Sts: en}
-					//if tag, ts, vals, e := storage.ReadSeriesTS(s0a, s1a, params["analysis"] != "current"); e == nil {
-					//	rta := s0.UnmarshalSliceSS(tag, ts, vals)
-					//	fmt.Println(rta)
-					//}
-					//fmt.Println(convertedRt)
-					//os.Exit(1)
-					// TODO development
+
+					fmt.Println(len(convertedRt))
+					for _, el := range convertedRt {
+						fmt.Println(el.Sts, el.Sval)
+					}
+					os.Exit(1)
+					// TODO end development
 					if e := json.NewEncoder(w).Encode(convertedRt); e != nil {
 						_, _ = fmt.Fprintf(w, "")
 					}